refactor: isolate workspace write policy
